server >> shared
