+++ Content.Server/_Mono/CorticalBorer/CorticalBorerSystem.cs
@@ -16,7 +16,6 @@ using Content.Shared.Chat; // Einstein Engines - Languages
 using Content.Shared.Chemistry.Components;
 using Content.Shared.Chemistry.Reagent;
 using Content.Shared.Database;
-using Content.Shared.Interaction.Events;
 using Content.Shared.Inventory;
 using Content.Shared.MedicalScanner;
 using Content.Shared.Mind;
@@ -51,6 +50,8 @@ public sealed partial class CorticalBorerSystem : SharedCorticalBorerSystem
 
     public override void Initialize()
     {
+        base.Initialize();
+
         SubscribeAbilities();
 
         SubscribeLocalEvent<CorticalBorerComponent, ComponentStartup>(OnStartup);
@@ -62,7 +63,6 @@ public sealed partial class CorticalBorerSystem : SharedCorticalBorerSystem
         SubscribeLocalEvent<CorticalBorerComponent, CheckTargetedSpeechEvent>(OnSpeakEvent);
 
         SubscribeLocalEvent<CorticalBorerComponent, MindRemovedMessage>(OnMindRemoved);
-        SubscribeLocalEvent<CorticalBorerComponent, AttackAttemptEvent>(OnBorerAttackAttempt);
     }
 
     private void OnStartup(Entity<CorticalBorerComponent> ent, ref ComponentStartup args)
@@ -189,22 +189,7 @@ public sealed partial class CorticalBorerSystem : SharedCorticalBorerSystem
         return true;
     }
 
-    /// <summary>
-    /// Checks if host has borer protection before allowing attack damage vs hosts
-    /// </summary>
-    private void OnBorerAttackAttempt(EntityUid uid, CorticalBorerComponent component, AttackAttemptEvent args)
-    {
-        if (args.Cancelled)
-            return;
 
-        if (component.Host is EntityUid host && args.Target == host)
-        {
-            if (!CanUseAbility((uid, component), host))
-            {
-                args.Cancel();
-            }
-        }
-    }
 
     private void OnInjectReagentMessage(Entity<CorticalBorerComponent> ent, ref CorticalBorerDispenserInjectMessage message)
     {
