+++ Content.Shared/_Mono/CorticalBorer/SharedCorticalBorerSystem.cs
@@ -1,23 +1,26 @@
-using System.Linq;
 using Content.Shared.Actions;
 using Content.Shared.Body.Components;
 using Content.Shared.Body.Part;
 using Content.Shared.Body.Systems;
 using Content.Shared.Containers.ItemSlots;
-using Content.Shared.Examine;
-using Content.Shared.MedicalScanner;
-using Content.Shared.Popups;
-using Content.Shared.StatusEffect;
 using Content.Shared.Coordinates;
 using Content.Shared.Damage;
+using Content.Shared.Examine;
 using Content.Shared.IdentityManagement;
+using Content.Shared.Interaction.Events;
+using Content.Shared.Inventory;
+using Content.Shared.MedicalScanner;
+using Content.Shared.Mind.Components;
+using Content.Shared.Popups;
+using Content.Shared.StatusEffect;
 using Robust.Shared.Containers;
 using Robust.Shared.Serialization;
 using Robust.Shared.Serialization.Manager;
+using System.Linq;
 
 namespace Content.Shared._Mono.CorticalBorer;
 
-public partial class SharedCorticalBorerSystem : EntitySystem
+public abstract class SharedCorticalBorerSystem : EntitySystem
 {
     [Dependency] private readonly SharedBodySystem _bodySystem = default!;
     [Dependency] private readonly StatusEffectsSystem _statusEffects = default!;
@@ -29,6 +32,11 @@ public partial class SharedCorticalBorerSystem : EntitySystem
     [Dependency] protected readonly SharedActionsSystem _actions = default!;
     [Dependency] protected readonly SharedContainerSystem _container = default!;
 
+    public override void Initialize()
+    {
+        SubscribeLocalEvent<CorticalBorerComponent, AttackAttemptEvent>(OnBorerAttackAttempt);
+    }
+
     public bool CanUseAbility(Entity<CorticalBorerComponent> ent, EntityUid target)
     {
         if (_statusEffects.HasStatusEffect(target,
@@ -39,6 +47,20 @@ public partial class SharedCorticalBorerSystem : EntitySystem
         }
 
         return true;
+    }
+
+    /// <summary>
+    /// Checks if host has borer protection before allowing attack damage vs hosts
+    /// </summary>
+    private void OnBorerAttackAttempt(EntityUid uid, CorticalBorerComponent component, AttackAttemptEvent args)
+    {
+        if (component.Host is EntityUid host && args.Target == host)
+        {
+            if (!CanUseAbility((uid, component), host))
+            {
+                args.Cancel();
+            }
+        }
     }
 
     public void InfestTarget(Entity<CorticalBorerComponent> ent, EntityUid target)
